changing the way to build dictionary like KOMORAN

+++ core/src/main/java/da/klay/core/morphology/analysis/sequence/AbstractMorphSequence.java
@@ -35,9 +35,9 @@ public abstract class AbstractMorphSequence implements MorphSequence {
     @Override
     public void compareScoreAndSetPreviousMSeq(MorphSequence newPreviousMSeq,
                                                TransitionMapBaseDictionary dictionary) {
-        Map<CharSequence, Integer> transitionMap = dictionary.getFully(newPreviousMSeq.last().getPos());
-        Integer transitionScore;
-        if(transitionMap == null || (transitionScore = transitionMap.get(first().getPos())) == null) transitionScore = -1;
+        Map<CharSequence, Double> transitionMap = dictionary.getFully(newPreviousMSeq.last().getPos());
+        Double transitionScore;
+        if(transitionMap == null || (transitionScore = transitionMap.get(first().getPos())) == null) transitionScore = -1.0;
 
         double newTotalScore = newPreviousMSeq.score() + transitionScore + emissionScore;
         if(newTotalScore < score && hasHPreviousMSeq()) return;
